add color grading

+++ shaders/customization/chair_ps.hlsl
@@ -19,7 +19,7 @@ half4 ps(PSInput input) : SV_TARGET
 	float4 albedo = lerp(lerp(ColoBase, ColorSeats, params.b), ColorMetal,params.g) * params.r;
 
 	half4 normal = calNormal(normalMap, linearSampler, input.uv,input.normal, input.tangent, input.binormal);
-	float roughness = lerp(lerp(0.5, 0.2, params.b), 0.4, params.g);
+	half roughness = lerp(lerp(0.5, 0.2, params.b), 0.4, params.g);
 
 	half3 color = directBRDF(roughness, params.g, F0_DEFAULT, albedo, normal,-sundir, campos - input.worldPos) ;
